siteIndex.json: rework how h1 is captured (#143)
Originally, we capture h1 elements by splitting on `>`, then
looking up the second element (that is, element 1) of the resulting
array for the h1 content. This does not work well when there are
elements preceding the h1 tag (eg, style elements)

Make the h1 capture logic more resilient by splitting instead on
`</h1>`, taking the first element of the split array, then splitting
on `">`. This assumes that all header elements will have an attribute
set (usually id). The header text is then found at the last element of
this array